Split up cleveldb
