+++ cleveldb.go
@@ -1,0 +1,176 @@
+// +build cleveldb
+
+package db
+
+import (
+	"fmt"
+	"path/filepath"
+
+	"github.com/jmhodges/levigo"
+)
+
+func init() {
+	dbCreator := func(name string, dir string) (DB, error) {
+		return NewCLevelDB(name, dir)
+	}
+	registerDBCreator(CLevelDBBackend, dbCreator, false)
+}
+
+// CLevelDB uses the C LevelDB database via a Go wrapper.
+type CLevelDB struct {
+	db     *levigo.DB
+	ro     *levigo.ReadOptions
+	wo     *levigo.WriteOptions
+	woSync *levigo.WriteOptions
+}
+
+var _ DB = (*CLevelDB)(nil)
+
+// NewCLevelDB creates a new CLevelDB.
+func NewCLevelDB(name string, dir string) (*CLevelDB, error) {
+	dbPath := filepath.Join(dir, name+".db")
+
+	opts := levigo.NewOptions()
+	opts.SetCache(levigo.NewLRUCache(1 << 30))
+	opts.SetCreateIfMissing(true)
+	db, err := levigo.Open(dbPath, opts)
+	if err != nil {
+		return nil, err
+	}
+	ro := levigo.NewReadOptions()
+	wo := levigo.NewWriteOptions()
+	woSync := levigo.NewWriteOptions()
+	woSync.SetSync(true)
+	database := &CLevelDB{
+		db:     db,
+		ro:     ro,
+		wo:     wo,
+		woSync: woSync,
+	}
+	return database, nil
+}
+
+// Get implements DB.
+func (db *CLevelDB) Get(key []byte) ([]byte, error) {
+	key = nonNilBytes(key)
+	res, err := db.db.Get(db.ro, key)
+	if err != nil {
+		return nil, err
+	}
+	return res, nil
+}
+
+// Has implements DB.
+func (db *CLevelDB) Has(key []byte) (bool, error) {
+	bytes, err := db.Get(key)
+	if err != nil {
+		return false, err
+	}
+	return bytes != nil, nil
+}
+
+// Set implements DB.
+func (db *CLevelDB) Set(key []byte, value []byte) error {
+	key = nonNilBytes(key)
+	value = nonNilBytes(value)
+	if err := db.db.Put(db.wo, key, value); err != nil {
+		return err
+	}
+	return nil
+}
+
+// SetSync implements DB.
+func (db *CLevelDB) SetSync(key []byte, value []byte) error {
+	key = nonNilBytes(key)
+	value = nonNilBytes(value)
+	if err := db.db.Put(db.woSync, key, value); err != nil {
+		return err
+	}
+	return nil
+}
+
+// Delete implements DB.
+func (db *CLevelDB) Delete(key []byte) error {
+	key = nonNilBytes(key)
+	if err := db.db.Delete(db.wo, key); err != nil {
+		return err
+	}
+	return nil
+}
+
+// DeleteSync implements DB.
+func (db *CLevelDB) DeleteSync(key []byte) error {
+	key = nonNilBytes(key)
+	if err := db.db.Delete(db.woSync, key); err != nil {
+		return err
+	}
+	return nil
+}
+
+// FIXME This should not be exposed
+func (db *CLevelDB) DB() *levigo.DB {
+	return db.db
+}
+
+// Close implements DB.
+func (db *CLevelDB) Close() error {
+	db.db.Close()
+	db.ro.Close()
+	db.wo.Close()
+	db.woSync.Close()
+	return nil
+}
+
+// Print implements DB.
+func (db *CLevelDB) Print() error {
+	itr, err := db.Iterator(nil, nil)
+	if err != nil {
+		return err
+	}
+	defer itr.Close()
+	for ; itr.Valid(); itr.Next() {
+		key := itr.Key()
+		value := itr.Value()
+		fmt.Printf("[%X]:\t[%X]\n", key, value)
+	}
+	return nil
+}
+
+// Stats implements DB.
+func (db *CLevelDB) Stats() map[string]string {
+	keys := []string{
+		"leveldb.aliveiters",
+		"leveldb.alivesnaps",
+		"leveldb.blockpool",
+		"leveldb.cachedblock",
+		"leveldb.num-files-at-level{n}",
+		"leveldb.openedtables",
+		"leveldb.sstables",
+		"leveldb.stats",
+	}
+
+	stats := make(map[string]string, len(keys))
+	for _, key := range keys {
+		str := db.db.PropertyValue(key)
+		stats[key] = str
+	}
+	return stats
+}
+
+// NewBatch implements DB.
+func (db *CLevelDB) NewBatch() Batch {
+	batch := levigo.NewWriteBatch()
+	return &cLevelDBBatch{db, batch}
+}
+
+// Iterator implements DB.
+func (db *CLevelDB) Iterator(start, end []byte) (Iterator, error) {
+	itr := db.db.NewIterator(db.ro)
+	return newCLevelDBIterator(itr, start, end, false), nil
+}
+
+// ReverseIterator implements DB.
+func (db *CLevelDB) ReverseIterator(start, end []byte) (Iterator, error) {
+	itr := db.db.NewIterator(db.ro)
+	return newCLevelDBIterator(itr, start, end, true), nil
+}
